WhizJuniors/requirements.txt
streamlit==1.24.0
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
plotly==5.15.0
